ai_service/requirements.txt
flask
flask-cors
werkzeug
pydub
openai-whisper @ git+https://github.com/openai/whisper.git
transformers
torch
gunicorn
